test : 위시 등록/취소 기능 통합 테스트 생성
- 회원이 작성한 쉐어에 대한 위시 등록 실패케이스 추가 구현

+++ src/main/java/louie/hanse/shareplate/service/WishService.java
@@ -4,6 +4,8 @@ import lombok.RequiredArgsConstructor;
 import louie.hanse.shareplate.domain.Member;
 import louie.hanse.shareplate.domain.Share;
 import louie.hanse.shareplate.domain.Wish;
+import louie.hanse.shareplate.exception.GlobalException;
+import louie.hanse.shareplate.exception.type.WishExceptionType;
 import louie.hanse.shareplate.repository.ShareRepository;
 import louie.hanse.shareplate.repository.WishRepository;
 import org.springframework.stereotype.Service;
@@ -15,13 +17,22 @@ import org.springframework.transaction.annotation.Transactional;
 public class WishService {
 
     private final MemberService memberService;
+    private final ShareService shareService;
     private final WishRepository wishRepository;
     private final ShareRepository shareRepository;
 
     @Transactional
     public void register(Long memberId, Long shareId) {
         Member member = memberService.findByIdOrElseThrow(memberId);
-        Share share = shareRepository.findById(shareId).orElseThrow();
+        Share share = shareService.findByIdOrElseThrow(shareId);
+
+        if (isExistWish(memberId, shareId)) {
+            throw new GlobalException(WishExceptionType.SHARE_ALREADY_WISH);
+        }
+
+        if (isWriter(memberId,shareId)) {
+            throw new GlobalException(WishExceptionType.WRITER_CAN_NOT_WISH);
+        }
 
         Wish wish = new Wish(share, member);
         wishRepository.save(wish);
@@ -29,6 +40,19 @@ public class WishService {
 
     @Transactional
     public void delete(Long memberId, Long shareId) {
+        shareService.findByIdOrElseThrow(shareId);
+        if (!isExistWish(memberId, shareId)) {
+            throw new GlobalException(WishExceptionType.SHARE_NOT_WISH);
+        }
+
         wishRepository.deleteByMemberIdAndShareId(memberId, shareId);
+    }
+
+    private boolean isExistWish(Long memberId, Long shareId) {
+        return wishRepository.existsByMemberIdAndShareId(memberId, shareId);
+    }
+
+    private boolean isWriter(Long memberId, Long shareId) {
+        return shareRepository.existsByIdAndWriterId(shareId, memberId);
     }
 }
